moved utility functions to separate utils.py
