+++ multipoint/__init__.py
@@ -3,7 +3,7 @@ __version__ = "1.2.0"
 
 from .multiPoint import multiPoint
 from .multiPointSparse import multiPointSparse
-from .multiPointSparse import createGroups
-from .multiPointSparse import redirectIO
+from .utils import createGroups
+from .utils import redirectIO
 
 __all__ = ["multiPoint", "multiPointSparse", "createGroups", "redirectIO"]
